+++ wizer/best_sections/climb.py
@@ -3,7 +3,7 @@ from sportgems import find_best_climb_section
 from wizer.best_sections.generic import GenericBestSection
 
 
-def get_best_climb_section(section_distance: int, parser) -> GenericBestSection:
+def get_best_climb_section(distance: int, parser) -> GenericBestSection:
     coordinates = list(zip(parser.latitude_list, parser.longitude_list))
 
     # safety check to catch activities with empty altitude list
@@ -12,7 +12,7 @@ def get_best_climb_section(section_distance: int, parser) -> GenericBestSection:
 
     # call to rust
     section = find_best_climb_section(
-        desired_distance=section_distance,
+        desired_distance=distance,
         times=parser.timestamps_list,
         coordinates=coordinates,
         altitudes=parser.altitude_list,
@@ -21,7 +21,7 @@ def get_best_climb_section(section_distance: int, parser) -> GenericBestSection:
     return GenericBestSection(
         start=section.start,
         end=section.end,
-        distance=section_distance,
+        distance=distance,
         max_value=round(section.climb, 2),
         kind="climb",
     )