+++ wizer/templatetags/filters.py
@@ -79,7 +79,7 @@ def queryset_to_list(queryset: QuerySet):
 def get_best_sections_list(queryset: QuerySet):
     list_of_start_and_end_index = []
     for section in queryset:
-        list_of_start_and_end_index.append([section.start_index, section.end_index])
+        list_of_start_and_end_index.append([section.start, section.end])
     return list_of_start_and_end_index